+++ src/C Files/employee_management.c
@@ -6,10 +6,28 @@ Option (3) to delete any employee record, (4) to view all the existing employee
 records, (5) to return back to main menu. So all the management of employee record
 is done in this file */
 
-/* Stucture definition for employee details to add and read an employee */
+/*
+ * @file employee_management.c
+ *
+ * @brief Has the primary functions of employee management..
+ *
+ * @author Praneeth Doppalapudi     - praneethdoppalapudi@cmail.carleton.ca
+ * @author Bhavani Singetam         - bsing105@uottawa.ca
+ * @author Sravya Yarlagadda        - sravyayarlagaddda@cmail.carleton.ca
+ * @author Hitesh Chowdary Nagalla  - hiteshchowdarynagall@cmail.carleton.ca
+ */
+
+
 
 #include<stdio.h>
-
+#include<conio.h>
+#include<string.h>
+#include<stdlib.h>
+#include<ctype.h>
+#include<windows.h>
+#include<time.h>
+#include<unistd.h>
+/* Stucture definition for employee details to add and read an employee */
 struct employee_details{
     char name[30];
     int id;
@@ -46,7 +64,15 @@ void employee_management(void){
     printf("\n\n");
     printf("\n\n");
 	if(i==1){
-		add_record(); /* function call to add a record*/
+		add_record();
+		printf("Save any more?(Y / N): ");
+		if(getch()=='n'){
+	    	employee_management();
+		}
+		else{
+	    	add_record();
+		}/* function call to add a record*/
+	employee_management();
 	}
 	else if(i==2){
 		search_record();/* function call to search a record*/
@@ -70,6 +96,7 @@ void employee_management(void){
 /*function to add a record*/
 void add_record(void){
 	system("cls");
+
 	if(getdata()==1){  /* function call to enter all the details*/
         fp=fopen("record1.txt","a+");
 		fseek(fp,0,SEEK_END);
@@ -77,71 +104,12 @@ void add_record(void){
 		fclose(fp);
 		printf("\n\n");
 		printf("The Record Is Successfully Saved ! !\n\n");
-		printf("Save any more?(Y / N): ");
-		if(getch()=='n'){
-	    	employee_management();
-		}
-		else{
-	    	add_record();
-		}
-	}else{
-	    employee_management();
 	}
-}
-
-
-	int checkid(int c){  /*check whether the record is exist in list or not*/
-	fp=fopen("record1.txt","a+");
-    fseek (fp, 0, SEEK_END);
-    if( ftell(fp) == 0){
-        return 1;  /*returns 1 if employee_details exits*/
-    }
-    else{
-        fseek (fp, 0, SEEK_SET);
-        while(fread(&e,sizeof(e),1,fp)==1){
-            if(e.id==c){
-                return 0;  /*returns 0 if employee_details exits*/
-            }
-            else{
-                return 1; /*return 1 if it not*/
-            }
-        }
-    }
-
-}
-
-/* function to add all employee details taken from user*/
-int getdata(){
-	int t = 0;
-	int  tempvar;
-	printf("\tEnter Employee ID :\t");
-	scanf("%d",&t);
-	fflush(stdin);
-	tempvar = checkid(t);
-	fclose(fp);
-	if(tempvar == 0){
-		printf("\n\n");
-		printf("\a***************The Employee Record Already Exists !!!\a");
-		getch();
-		return 0;
-	}else if(tempvar==1){
-        e.id=t;
-        printf("\tEnter Employee Name : ");
-        scanf("%s",e.name);
-        fflush(stdin);
-        printf("        Date Of Birth (dd/mm/yyyy) :");
-        scanf("%d/%d/%d",&e.dd,&e.mm,&e.yyyy);
-        fflush(stdin);
-        printf("        Year OF Joining :");
-        scanf("%d",&e.YOJ);
-        fflush(stdin);
-        printf("        Place :");
-        scanf("%s",e.place);
-        fflush(stdin);
-        printf("        Department :");
-        scanf("%s",e.department);
-        fflush(stdin);
-        return 1;
+	else if(getdata()==-1){
+		return;
+	}
+	else{
+          employee_management();
 	}
 }
 
@@ -154,7 +122,7 @@ void search_record(void){
     printf("                  2. Search By Name\n\n");
     printf("                  Enter Your Choice---> ");
     fp=fopen("record1.txt","r"); /*open file for reading propose*/
-    rewind(fp);   /*move pointer at the begining of file*/
+    fseek (fp, 0, SEEK_SET);   /*move pointer at the begining of file*/
     switch(getch()){
         case '1':{ /*search recorde by id*/
 	    	system("cls");
@@ -234,54 +202,54 @@ void delete_record(void){
     int d;
     char another='y';
     while(another=='y'){
-	system("cls");
-	printf("                Enter The Employee ID To Delete :");
-	scanf(" %d",&d);
-	fp =fopen("record1.txt","r+");
-	rewind(fp);
-	while(fread(&e,sizeof(e),1,fp)==1){
-	    if(e.id==d){
-			printf("\n\n");
-			printf(".....................The Employee Record Is Available....................\n\n");
-			printf("               Employee's Name Is %s\n\n",e.name);
-			printf("               Employee's Date  OF Birth Is %d/%d/%d\n\n",e.dd,e.mm,e.yyyy);
-			findrecord='t';
-	    }
-	}
-	if(findrecord!='t'){
-	    printf(".........................No record is found modify the search.................\n\n");
-	    if(getch())
-	    employee_management();
-	}
-	if(findrecord=='t' ){
-	    printf("Do You Want To Delete The Record ? (Y/N) ");
-
-	    if(getch()=='y'){
-	        ft=fopen("test1.txt","a+");  /*temporary file for delete*/
-                fseek (fp, 0, SEEK_SET);
-			while(fread(&e,sizeof(e),1,fp)==1){
-			    printf("%d %d",e.id,d);
-		    	if(e.id!=d){
-		    	    printf("%d %d",e.id,d);
-					fwrite(&e,sizeof(e),1,ft); /*write all in tempory file except that*/
-
-		    	}                              /*we want to delete*/
+		system("cls");
+		printf("                Enter The Employee ID To Delete :");
+		scanf(" %d",&d);
+		fp =fopen("record1.txt","r+");
+		rewind(fp);
+		while(fread(&e,sizeof(e),1,fp)==1){
+			if(e.id==d){
+				printf("\n\n");
+				printf(".....................The Employee Record Is Available....................\n\n");
+				printf("               Employee's Name Is %s\n\n",e.name);
+				printf("               Employee's Date  OF Birth Is %d/%d/%d\n\n",e.dd,e.mm,e.yyyy);
+				findrecord='t';
 			}
-        fclose(ft);
-        fclose(fp);
-			int q = remove("record1.txt");
-			int w = rename("test1.txt","record1.txt"); /*copy all item from temporary file to fp except that*/
-			printf("%d %d", q,w);
-            printf("        THE RECORD IS SUCCESSFULLY DELETED.\n\n");
-            printf("        Delete Another Record ? (Y/N) : ");
-	    }
-		else{
-		employee_management();
 		}
-		fflush(stdin);
-		another=getch();
+		if(findrecord!='t'){
+			printf(".........................No record is found modify the search.................\n\n");
+			if(getch()){
+				employee_management();
+			}
+		}
+		if(findrecord=='t' ){
+			printf("Do You Want To Delete The Record ? (Y/N) ");
+
+			if(getch()=='y'){
+				ft=fopen("test1.txt","a+");  /*temporary file for delete*/
+				fseek (fp, 0, SEEK_SET);
+				while(fread(&e,sizeof(e),1,fp)==1){
+					printf("%d %d",e.id,d);
+					if(e.id!=d){
+						printf("%d %d",e.id,d);
+						fwrite(&e,sizeof(e),1,ft); /*write all in tempory file except that*/
+					}
+				}
+				fclose(ft);
+				fclose(fp);
+				int q = remove("record1.txt");
+				int w = rename("test1.txt","record1.txt"); /*copy all item from temporary file to fp except that*/
+				printf("%d %d", q,w);
+				printf("        THE RECORD IS SUCCESSFULLY DELETED.\n\n");
+				printf("        Delete Another Record ? (Y/N) : ");
+			}
+			else{
+			employee_management();
+			}
+			fflush(stdin);
+			another=getch();
+		}
 	}
-}
 employee_management();
 }
 
@@ -307,6 +275,151 @@ void view_record(void){
 }
 
 
+int checkid(int c){  /*check whether the record is exist in list or not*/
+	fp=fopen("record1.txt","a+");
+	if(fp==0) {
+        fp=fopen("record1.txt","w+");
+		system("cls");
+		printf("Please hold on while we set our database in your computer!!");
+		printf("\n Process completed press any key to continue!! ");
+		getch();
+    }
+	fseek (fp, 0, SEEK_END);
+	if( ftell(fp) == 0){
+		return 1;  /*returns 1 if employee_details exits*/
+	}
+	else{
+		fseek (fp, 0, SEEK_SET);
+		while(fread(&e,sizeof(e),1,fp)==1){
+			if(e.id==c){
+				return 0;  /*returns 0 if employee_details exits*/
+			}
+			else{
+				return 1; /*return 1 if it not*/
+			}
+		}
+	}
+}
 
-
-
+/* function to add all employee details taken from user*/
+int getdata(){
+	int t = 0;
+	int  tempvar;
+	int checkdata = 0;
+	int count = 0;
+	printf("\tEnter Employee ID :\t");
+	scanf("%d",&t);
+	fflush(stdin);
+	tempvar = checkid(t);
+	fclose(fp);
+	if(tempvar == 0){
+		printf("\n\n");
+		printf("\a***************The Employee Record Already Exists !!!\a");
+		getch();
+		return 0;
+	}
+	else if(tempvar==1){
+        e.id=t;
+        printf("\tEnter Employee Name : ");
+        checkdata = scanf("%s",e.name);
+        fflush(stdin);
+        count = 0;
+        while(1){
+            if(count >=5){
+                return -1;
+            }
+            else{
+                if(!checkdata){
+                    printf("Enter Valid Input data! \n");
+                    checkdata = scanf("%s",e.name);
+                    fflush(stdin);
+                    count+=1;
+                }
+                else{
+                    break;
+                }
+            }
+        }
+        printf("\tDate Of Birth (dd/mm/yyyy) :");
+        checkdata = scanf("%d/%d/%d",&e.dd,&e.mm,&e.yyyy);
+        fflush(stdin);
+        count = 0;
+        while(1){
+            if(count >=5){
+                return -1;
+            }
+            else{
+                if(!checkdata){
+                    printf("Enter Valid Input data! \n");
+                    checkdata = scanf("%d/%d/%d",&e.dd,&e.mm,&e.yyyy);
+                    fflush(stdin);
+                    count+=1;
+                }
+                else{
+                    break;
+                }
+            }
+        }
+        printf("\tYear OF Joining :");
+        checkdata = scanf("%d",&e.YOJ);
+        fflush(stdin);
+        count = 0;
+        while(1){
+            if(count >=5){
+                return -1;
+            }
+            else{
+                if(!checkdata){
+                    printf("Enter Valid Input data! \n");
+                    checkdata = scanf("%d",&e.YOJ);
+                    fflush(stdin);
+                    count+=1;
+                }
+                else{
+                    break;
+                }
+            }
+        }
+        printf("\tPlace :");
+        checkdata = scanf("%s",e.place);
+        fflush(stdin);
+        count = 0;
+        while(1){
+            if(count >=5){
+                return -1;
+            }
+            else{
+                if(!checkdata){
+                    printf("Enter Valid Input data! \n");
+                    checkdata = scanf("%s",e.place);
+                    fflush(stdin);
+                    count+=1;
+                }
+                else{
+                    break;
+                }
+            }
+        }
+        printf("        Department :");
+        checkdata = scanf("%s",e.department);
+        fflush(stdin);
+        count = 0;
+        while(1){
+            if(count >=5){
+                return -1;
+            }
+            else{
+                if(!checkdata){
+                    printf("Enter Valid Input data! \n");
+                    checkdata = scanf("%s",e.department);
+                    fflush(stdin);
+                    count+=1;
+                }
+                else{
+                    break;
+                }
+            }
+        }
+        return 1;
+	}
+}
